chtml/html/testdata/go/issue_30600_parse_panics_in_cell_mode.dat

#data
<table><math><th><mo><select></table>
#errors
#document
| <math math>
|   <math th>
|     <math mo>
|       <select>
| <table>
